Separated test files from source files
