+++ test/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"compilerOptions": {
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext",
+		"target": "ES2022",
+		"skipLibCheck": true,
+		"composite": true
+	}
+}
