+++ tsconfig.json
@@ -6,5 +6,10 @@
 		"outDir": "aws_lambda",
 		"skipLibCheck": true
 	},
-	"include": ["src/**/*"]
+	"include": ["src/**/*"],
+	"references": [
+		{
+			"path": "./test"
+		}
+	]
 }
